[JAVA-26439] Moved code related to "Guide to Stream.reduce()" to ebook module
