--- core-java-modules/core-java-8/src/main/java/com/baeldung/streamreduce/entities/User.java
@@ -1,30 +1,0 @@
-package com.baeldung.streamreduce.entities;
-
-public class User {
-
-    private final String name;
-    private final int age;
-    private final Rating rating = new Rating();
-
-    public User(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public Rating getRating() {
-        return rating;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" + "name=" + name + ", age=" + age + '}';
-    }
-}
