+++ core-java-modules/core-java-streams-simple/src/main/java/com/baeldung/streams/reduce/entities/User.java
@@ -4,6 +4,7 @@ public class User {
 
     private final String name;
     private final int age;
+    private final Rating rating = new Rating();
 
     public User(String name, int age) {
         this.name = name;
@@ -17,7 +18,11 @@ public class User {
     public int getAge() {
         return age;
     }
-    
+
+    public Rating getRating() {
+        return rating;
+    }
+
     @Override
     public String toString() {
         return "User{" + "name=" + name + ", age=" + age + '}';
